tests pass with defaultTo / _defaultValue removed
